inventory-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic[email]==2.7.1
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.0
pytest-mock==3.14.0
